REFACTOR-002: extract clip invariants into ./parse/clip.ts
Pure structural move, no behavior change. Moves checkClipInvariants
(exported) and checkClipOwnership (module-private, used by
checkProjectIntegrity) out of parse.ts into their own module. parse.ts
re-exports checkClipInvariants unchanged, so no consumer import moves.

Second of the REFACTOR-002 stack; builds on the primitives extraction.

Refs #143
